perf: optimize bundle size with dynamic imports and code splitting
- Create AsyncPdfRenderer wrapper for dynamic loading of PDF components
- Split UploadDropzone from UploadButton for dashboard optimization
- Dynamically import Messages in ChatWrapper
- Dynamically import PdfRenderer in dashboard and demo pages
