+++ src/app/dashboard/[fileid]/page.tsx
@@ -1,9 +1,9 @@
 import { ChatWrapper } from "@/components/chat/ChatWrapper";
-import { PdfRenderer } from "@/components/PdfRenderer";
 import { db } from "@/lib/db";
 import { requireUser } from "@/lib/auth";
 import { notFound } from "next/navigation";
 import { Metadata, Viewport } from "next";
+import AsyncPdfRenderer from "@/components/AsyncPdfRenderer";
 
 interface PageProps {
   params: {
@@ -44,7 +44,7 @@ const Page = async ({ params }: PageProps) => {
         {/* left side */}
         <div className="flex-1 xl:flex">
           <div className="px-4 py-6 sm:px-6 lg:pl-8 xl:flex-1 xl:pl-6">
-            <PdfRenderer url={file.url} />
+            <AsyncPdfRenderer url={file.url} />
           </div>
         </div>
